@@ -1,13 +1,13 @@
 buildscript {
     ext {
-        kotlin_version = '1.7.10'
+        kotlin_version = '1.8.22'
         coroutinesVersion = '1.6.4'
         ok_http_version = '4.9.1'
         dashjVersion = '19.0.3'
-        hiltVersion = '2.40.1'
+        hiltVersion = '2.45'
         hiltWorkVersion = '1.0.0'
         workRuntimeVersion='2.7.1'
-        firebaseVersion = '28.4.2'
+        firebaseVersion = '32.1.1'
         roomVersion = '2.4.3'
         jetpackVersion = '1.8.0'
         appCompatVersion = '1.4.2'
