@@ -32,6 +32,7 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AlertDialog
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.constraintlayout.widget.ConstraintSet
+import androidx.core.view.isVisible
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.lifecycleScope
@@ -42,16 +43,15 @@ import de.schildbach.wallet.WalletApplication
 import de.schildbach.wallet.livedata.Status
 import de.schildbach.wallet.security.BiometricHelper
 import de.schildbach.wallet.security.BiometricLockoutException
+import de.schildbach.wallet.security.PinRetryController
 import de.schildbach.wallet.service.PackageInfoProvider
 import de.schildbach.wallet.service.RestartService
 import de.schildbach.wallet.ui.payments.QuickReceiveActivity
-import de.schildbach.wallet.security.PinRetryController
 import de.schildbach.wallet.ui.send.SendCoinsQrActivity
 import de.schildbach.wallet.ui.verify.VerifySeedActivity
 import de.schildbach.wallet.ui.widget.PinPreviewView
 import de.schildbach.wallet_test.R
-import kotlinx.android.synthetic.main.activity_lock_screen.*
-import kotlinx.android.synthetic.main.activity_lock_screen_root.*
+import de.schildbach.wallet_test.databinding.ActivityLockScreenRootBinding
 import kotlinx.coroutines.launch
 import org.bitcoinj.wallet.Wallet.BalanceType
 import org.dash.wallet.common.Configuration
@@ -85,10 +85,11 @@ open class LockScreenActivity : SecureActivity() {
 
     private val autoLogout: AutoLogout by lazy { walletApplication.autoLogout }
     private val checkPinViewModel by viewModels<CheckPinViewModel>()
+    private lateinit var binding: ActivityLockScreenRootBinding
     private val pinLength by lazy { configuration.pinLength }
 
     val lockScreenDisplayed: Boolean
-        get() = root_view_switcher.displayedChild == 0
+        get() = binding.rootViewSwitcher.displayedChild == 0
 
     private val temporaryLockCheckHandler = Handler()
     private val temporaryLockCheckInterval = TimeUnit.SECONDS.toMillis(10)
@@ -124,7 +125,8 @@ open class LockScreenActivity : SecureActivity() {
             return
         }
 
-        super.setContentView(R.layout.activity_lock_screen_root)
+        binding = ActivityLockScreenRootBinding.inflate(layoutInflater)
+        super.setContentView(binding.root)
         setupKeyboardBottomMargin()
 
         initView()
@@ -138,8 +140,8 @@ open class LockScreenActivity : SecureActivity() {
     }
 
     override fun setContentView(contentView: View?) {
-        regular_content.removeAllViews()
-        regular_content.addView(contentView)
+        binding.regularContent.removeAllViews()
+        binding.regularContent.addView(contentView)
     }
 
     private val onLogoutListener = AutoLogout.OnLogoutListener {
@@ -157,16 +159,6 @@ open class LockScreenActivity : SecureActivity() {
         }
     }
 
-    protected open fun turnOffAutoLogout() {
-        autoLogout.stopTimer()
-    }
-
-    protected open fun turnOnAutoLogout() {
-        if (!autoLogout.isTimerActive) {
-            autoLogout.startTimer()
-        }
-    }
-
     private fun resetAutoLogoutTimer() {
         autoLogout.resetTimerIfActive()
     }
@@ -181,7 +173,7 @@ open class LockScreenActivity : SecureActivity() {
     private fun setupKeyboardBottomMargin() {
         if (!hasNavBar()) {
             val set = ConstraintSet()
-            val layout = numeric_keyboard.parent as ConstraintLayout
+            val layout = binding.lockScreen.numericKeyboard.parent as ConstraintLayout
             set.clone(layout)
             set.clear(R.id.numeric_keyboard, ConstraintSet.BOTTOM)
             set.connect(R.id.numeric_keyboard, ConstraintSet.BOTTOM, ConstraintSet.PARENT_ID, ConstraintSet.BOTTOM)
@@ -202,13 +194,14 @@ open class LockScreenActivity : SecureActivity() {
             "K5", "K5c", "K5b", "K4m", "KRIP_K4" -> true
             else -> false
         }
+
         return if (id > 0 && !isKripDeviceWithoutNavBar) {
-            id > 0 && resources.getBoolean(id)
+            resources.getBoolean(id)
         } else {
             // Check for keys
-            val hasMenuKey = ViewConfiguration.get(this).hasPermanentMenuKey();
-            val hasBackKey = KeyCharacterMap.deviceHasKey(KeyEvent.KEYCODE_BACK);
-            !hasMenuKey && !hasBackKey;
+            val hasMenuKey = ViewConfiguration.get(this).hasPermanentMenuKey()
+            val hasBackKey = KeyCharacterMap.deviceHasKey(KeyEvent.KEYCODE_BACK)
+            !hasMenuKey && !hasBackKey
         }
     }
 
@@ -216,19 +209,23 @@ open class LockScreenActivity : SecureActivity() {
         super.onStart()
         autoLogout.setOnLogoutListener(onLogoutListener)
 
-        if (!keepUnlocked && configuration.autoLogoutEnabled && (autoLogout.keepLockedUntilPinEntered || autoLogout.shouldLogout())) {
-            setLockState(if (pinRetryController.isLocked) {
-                State.LOCKED
-            } else {
-                State.USE_DEFAULT
-            })
+        if (!keepUnlocked && configuration.autoLogoutEnabled &&
+            (autoLogout.keepLockedUntilPinEntered || autoLogout.shouldLogout())
+        ) {
+            setLockState(
+                if (pinRetryController.isLocked) {
+                    State.LOCKED
+                } else {
+                    State.USE_DEFAULT
+                }
+            )
             autoLogout.setAppWentBackground(false)
             if (autoLogout.isTimerActive) {
                 autoLogout.stopTimer()
             }
             handleLockScreenActivated()
         } else {
-            root_view_switcher.displayedChild = 1
+            binding.rootViewSwitcher.displayedChild = 1
             if (!keepUnlocked) {
                 autoLogout.maybeStartAutoLogoutTimer()
             }
@@ -242,50 +239,51 @@ open class LockScreenActivity : SecureActivity() {
     }
 
     private fun initView() {
-        action_login_with_pin.setOnClickListener {
-            setLockState(State.ENTER_PIN)
-        }
-        action_login_with_fingerprint.setOnClickListener {
-            setLockState(State.USE_FINGERPRINT)
-        }
-        action_receive.setOnClickListener {
-            startActivity(QuickReceiveActivity.createIntent(this))
-            autoLogout.keepLockedUntilPinEntered = true
-        }
-        action_scan_to_pay.setOnClickListener {
-            startActivity(SendCoinsQrActivity.createIntent(this, true))
-            autoLogout.keepLockedUntilPinEntered = true
-        }
-        numeric_keyboard.isFunctionEnabled = false
-        numeric_keyboard.onKeyboardActionListener = object : NumericKeyboardView.OnKeyboardActionListener {
+        binding.lockScreen.apply {
+            actionLoginWithPin.setOnClickListener {
+                setLockState(State.ENTER_PIN)
+            }
+            actionLoginWithFingerprint.setOnClickListener {
+                setLockState(State.USE_FINGERPRINT)
+            }
+            actionReceive.setOnClickListener {
+                startActivity(QuickReceiveActivity.createIntent(this@LockScreenActivity))
+                autoLogout.keepLockedUntilPinEntered = true
+            }
+            actionScanToPay.setOnClickListener {
+                startActivity(SendCoinsQrActivity.createIntent(this@LockScreenActivity, true))
+                autoLogout.keepLockedUntilPinEntered = true
+            }
+            numericKeyboard.isFunctionEnabled = false
+            numericKeyboard.onKeyboardActionListener = object : NumericKeyboardView.OnKeyboardActionListener {
 
-            override fun onNumber(number: Int) {
-                if (pinRetryController.isLocked) {
-                    return
+                override fun onNumber(number: Int) {
+                    if (pinRetryController.isLocked) {
+                        return
+                    }
+                    if (checkPinViewModel.pin.length < pinLength) {
+                        checkPinViewModel.pin.append(number)
+                        pinPreview.next()
+                    }
+                    if (checkPinViewModel.pin.length == pinLength) {
+                        pinPreview.postDelayed({
+                            checkPinViewModel.checkPin(checkPinViewModel.pin)
+                        }, 200)
+                    }
                 }
-                if (checkPinViewModel.pin.length < pinLength) {
-                    checkPinViewModel.pin.append(number)
-                    pin_preview.next()
+
+                override fun onBack(longClick: Boolean) {
+                    if (checkPinViewModel.pin.isNotEmpty()) {
+                        checkPinViewModel.pin.deleteCharAt(checkPinViewModel.pin.length - 1)
+                        pinPreview.prev()
+                    }
                 }
-                if (checkPinViewModel.pin.length == pinLength) {
-                    Handler().postDelayed({
-                        checkPinViewModel.checkPin(checkPinViewModel.pin)
-                    }, 200)
+
+                override fun onFunction() {
                 }
             }
-
-            override fun onBack(longClick: Boolean) {
-                if (checkPinViewModel.pin.isNotEmpty()) {
-                    checkPinViewModel.pin.deleteCharAt(checkPinViewModel.pin.length - 1)
-                    pin_preview.prev()
-                }
-            }
-
-            override fun onFunction() {
-
-            }
+            viewFlipper.inAnimation = AnimationUtils.loadAnimation(this@LockScreenActivity, android.R.anim.fade_in)
         }
-        view_flipper.inAnimation = AnimationUtils.loadAnimation(this, android.R.anim.fade_in)
     }
 
     private fun initViewModel() {
@@ -331,108 +329,108 @@ open class LockScreenActivity : SecureActivity() {
             finish()
             overridePendingTransition(android.R.anim.fade_in, android.R.anim.fade_out)
         } else {
-            root_view_switcher.displayedChild = 1
+            binding.rootViewSwitcher.displayedChild = 1
         }
 
         onLockScreenDeactivated()
     }
 
     private fun setLockState(suggestedState: State) {
-        action_scan_to_pay.isEnabled = true
-        var lockState = suggestedState
+        binding.lockScreen.apply {
+            actionScanToPay.isEnabled = true
+            var lockState = suggestedState
 
-        if (lockState == State.USE_DEFAULT) {
-            lockState = if (biometricHelper.isEnabled) {
-                State.USE_FINGERPRINT
-            } else {
-                log.info("fingerprint was disabled")
-                action_login_with_fingerprint.isEnabled = false
-                action_login_with_fingerprint.alpha = 0f
-                State.ENTER_PIN
-            }
-        }
-
-        log.info("LockState = $lockState")
-
-        when (lockState) {
-            State.ENTER_PIN, State.INVALID_PIN -> {
-                if (pinLength != PinPreviewView.DEFAULT_PIN_LENGTH) {
-                    pin_preview.mode = PinPreviewView.PinType.CUSTOM
-                }
-                view_flipper.displayedChild = 0
-
-                action_title.setText(R.string.lock_enter_pin)
-
-                action_login_with_pin.visibility = View.GONE
-                action_login_with_fingerprint.visibility = View.VISIBLE
-
-                numeric_keyboard.visibility = View.VISIBLE
-
-                if (lockState == State.INVALID_PIN) {
-                    checkPinViewModel.pin.clear()
-                    pin_preview.shake()
-                    Handler().postDelayed({
-                        pin_preview.clear()
-                    }, 200)
+            if (lockState == State.USE_DEFAULT) {
+                lockState = if (biometricHelper.isEnabled) {
+                    State.USE_FINGERPRINT
                 } else {
-                    numeric_keyboard.isEnabled = true
-                    pin_preview.clear()
+                    log.info("fingerprint was disabled")
+                    actionLoginWithFingerprint.isEnabled = false
+                    actionLoginWithFingerprint.alpha = 0f
+                    State.ENTER_PIN
+                }
+            }
+
+            log.info("LockState = $lockState")
+
+            when (lockState) {
+                State.ENTER_PIN, State.INVALID_PIN -> {
+                    if (pinLength != PinPreviewView.DEFAULT_PIN_LENGTH) {
+                        pinPreview.mode = PinPreviewView.PinType.CUSTOM
+                    }
+
+                    viewFlipper.displayedChild = 0
+                    actionTitle.setText(R.string.lock_enter_pin)
+                    actionLoginWithPin.isVisible = false
+                    actionLoginWithFingerprint.isVisible = true
+                    numericKeyboard.isVisible = true
+
+                    if (lockState == State.INVALID_PIN) {
+                        checkPinViewModel.pin.clear()
+                        pinPreview.shake()
+                        pinPreview.postDelayed({
+                            pinPreview.clear()
+                        }, 200)
+                    } else {
+                        numericKeyboard.isEnabled = true
+                        pinPreview.clear()
+                        checkPinViewModel.pin.clear()
+                        pinPreview.clearBadPin()
+                    }
+
+                    if (pinRetryController.failCount() > 0) {
+                        pinPreview.badPin(pinRetryController.getRemainingAttemptsMessage(resources))
+                    }
+
+                    if (pinRetryController.remainingAttempts == 1) {
+                        val dialog = AdaptiveDialog.create(
+                            R.drawable.ic_warning,
+                            getString(R.string.wallet_last_attempt),
+                            getString(R.string.wallet_last_attempt_message),
+                            "",
+                            getString(R.string.button_understand)
+                        )
+                        dialog.isCancelable = false
+                        dialog.show(this@LockScreenActivity)
+                    }
+                }
+                State.USE_FINGERPRINT -> {
+                    viewFlipper.displayedChild = 1
+                    actionTitle.setText(R.string.lock_unlock_with_fingerprint)
+                    actionLoginWithPin.visibility = View.VISIBLE
+                    actionLoginWithFingerprint.visibility = View.GONE
+                    numericKeyboard.visibility = View.GONE
+
+                    showBiometricPrompt()
+                }
+                State.DECRYPTING -> {
+                    viewFlipper.displayedChild = 2
+                    numericKeyboard.isEnabled = false
+                }
+                State.LOCKED -> {
+                    viewFlipper.displayedChild = 3
                     checkPinViewModel.pin.clear()
-                    pin_preview.clearBadPin()
+                    pinPreview.clear()
+                    temporaryLockCheckHandler.postDelayed(temporaryLockCheckRunnable, temporaryLockCheckInterval)
+
+                    actionTitle.setText(R.string.wallet_lock_wallet_disabled)
+                    actionSubtitle.text = pinRetryController.getWalletTemporaryLockedMessage(resources)
+
+                    actionLoginWithPin.visibility = View.GONE
+                    actionLoginWithFingerprint.visibility = View.GONE
+
+                    actionScanToPay.isEnabled = false
+                    numericKeyboard.visibility = View.GONE
                 }
-
-                if (pinRetryController.failCount() > 0) {
-                    pin_preview.badPin(pinRetryController.getRemainingAttemptsMessage(resources))
+                else -> {
+                    // we should never reach this since default means we use
+                    // ENTER_PIN or USE_FINGERPRINT
                 }
-
-                if (pinRetryController.remainingAttempts == 1) {
-                    val dialog = AdaptiveDialog.create(
-                        R.drawable.ic_error,
-                        getString(R.string.wallet_last_attempt),
-                        getString(R.string.wallet_last_attempt_message),
-                        "",
-                        getString(R.string.button_understand)
-                    )
-                    dialog.isCancelable = false
-                    dialog.show(this) { }
-                }
-            }
-            State.USE_FINGERPRINT -> {
-                view_flipper.displayedChild = 1
-                action_title.setText(R.string.lock_unlock_with_fingerprint)
-                action_login_with_pin.visibility = View.VISIBLE
-                action_login_with_fingerprint.visibility = View.GONE
-                numeric_keyboard.visibility = View.GONE
-
-                showBiometricPrompt()
-            }
-            State.DECRYPTING -> {
-                view_flipper.displayedChild = 2
-                numeric_keyboard.isEnabled = false
-            }
-            State.LOCKED -> {
-                view_flipper.displayedChild = 3
-                checkPinViewModel.pin.clear()
-                pin_preview.clear()
-                temporaryLockCheckHandler.postDelayed(temporaryLockCheckRunnable, temporaryLockCheckInterval)
-
-                action_title.setText(R.string.wallet_lock_wallet_disabled)
-                action_subtitle.text = pinRetryController.getWalletTemporaryLockedMessage(resources)
-
-                action_login_with_pin.visibility = View.GONE
-                action_login_with_fingerprint.visibility = View.GONE
-
-                action_scan_to_pay.isEnabled = false
-                numeric_keyboard.visibility = View.GONE
-            }
-            else -> {
-                // we should never reach this since default means we use
-                // ENTER_PIN or USE_FINGERPRINT
             }
         }
 
         if (!lockScreenDisplayed) {
-            root_view_switcher.displayedChild = 0
+            binding.rootViewSwitcher.displayedChild = 0
         }
     }
 
@@ -445,13 +443,13 @@ open class LockScreenActivity : SecureActivity() {
 
                 if (error is KeyPermanentlyInvalidatedException) {
                     showFingerprintKeyChangedDialog()
-                    action_login_with_fingerprint.isEnabled = false
+                    binding.lockScreen.actionLoginWithFingerprint.isEnabled = false
                 } else {
-                    fingerprint_view.showError(exceededMaxAttempts)
+                    binding.lockScreen.fingerprintView.showError(exceededMaxAttempts)
                 }
             } else if (savedPass != null) {
                 log.info("fingerprint scan successful")
-                fingerprint_view.hideError()
+                binding.lockScreen.fingerprintView.hideError()
                 onCorrectPin(savedPass)
             } else {
                 log.info("user canceled fingerprint scan")
@@ -484,7 +482,7 @@ open class LockScreenActivity : SecureActivity() {
     }
 
     private fun handleLockScreenActivated() {
-        if (this::alertDialog.isInitialized){
+        if (this::alertDialog.isInitialized) {
             alertDialog.dismissDialog()
         }
         lockScreenBroadcaster.activatingLockScreen.call()
