@@ -24,23 +24,21 @@ import android.content.Intent
 import android.graphics.drawable.LayerDrawable
 import android.os.Build
 import android.os.Bundle
-import android.os.Handler
 import android.text.TextUtils
-import android.view.View
-import android.widget.LinearLayout
 import android.widget.Toast
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
+import androidx.core.view.isVisible
 import dagger.hilt.android.AndroidEntryPoint
 import de.schildbach.wallet.WalletApplication
-import de.schildbach.wallet.ui.backup.RestoreFromFileActivity
-import de.schildbach.wallet.ui.main.WalletActivity
 import de.schildbach.wallet.security.PinRetryController
 import de.schildbach.wallet.security.SecurityGuard
+import de.schildbach.wallet.ui.backup.RestoreFromFileActivity
+import de.schildbach.wallet.ui.main.WalletActivity
 import de.schildbach.wallet_test.R
-import kotlinx.android.synthetic.main.activity_onboarding.*
-import kotlinx.android.synthetic.main.activity_onboarding_perm_lock.*
+import de.schildbach.wallet_test.databinding.ActivityOnboardingBinding
+import de.schildbach.wallet_test.databinding.ActivityOnboardingPermLockBinding
+import org.dash.wallet.common.Configuration
 import org.dash.wallet.common.services.analytics.AnalyticsService
 import org.dash.wallet.common.ui.dialogs.AdaptiveDialog
 import org.dash.wallet.common.util.getMainTask
@@ -87,34 +85,43 @@ class OnboardingActivity : RestoreFromFileActivity() {
         }
     }
 
-    private lateinit var viewModel: OnboardingViewModel
-
-    private lateinit var walletApplication: WalletApplication
+    private val viewModel by viewModels<OnboardingViewModel>()
+    private lateinit var binding: ActivityOnboardingBinding
 
     @Inject
+    lateinit var walletApplication: WalletApplication
+    @Inject
     lateinit var analytics: AnalyticsService
+    @Inject
+    lateinit var config: Configuration
+    @Inject
+    lateinit var pinRetryController: PinRetryController
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        walletApplication = (application as WalletApplication)
 
-        if (PinRetryController.getInstance().isLockedForever) {
-            setContentView(R.layout.activity_onboarding_perm_lock)
+        if (pinRetryController.isLockedForever) {
+            val binding = ActivityOnboardingPermLockBinding.inflate(layoutInflater)
+            setContentView(binding.root)
             getStatusBarHeightPx()
             hideSlogan()
-            close_app.setOnClickListener {
+            binding.closeApp.setOnClickListener {
                 finish()
             }
-            wipe_wallet.setOnClickListener {
+            binding.wipeWallet.setOnClickListener {
                 ResetWalletDialog.newInstance().show(supportFragmentManager, "reset_wallet_dialog")
             }
             return
         }
 
-        setContentView(R.layout.activity_onboarding)
-        slogan.setPadding(slogan.paddingLeft, slogan.paddingTop, slogan.paddingRight, getStatusBarHeightPx())
-
-        viewModel = ViewModelProvider(this)[OnboardingViewModel::class.java]
+        binding = ActivityOnboardingBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        binding.slogan.setPadding(
+            binding.slogan.paddingLeft,
+            binding.slogan.paddingTop,
+            binding.slogan.paddingRight,
+            getStatusBarHeightPx()
+        )
 
         // TODO: we should decouple the logic from view interactions
         // and move some of this to the viewModel, wrapping it in tests.
@@ -152,18 +159,23 @@ class OnboardingActivity : RestoreFromFileActivity() {
     // such that the wallet is not encrypted
     private fun unencryptedFlow() {
         log.info("the wallet is not encrypted -- the wallet will be upgraded")
-        if (walletApplication.configuration.v7TutorialCompleted) {
+        if (config.v7TutorialCompleted) {
             upgradeUnencryptedWallet()
         } else {
-            startActivityForResult(Intent(this, WelcomeActivity::class.java),
-                UPGRADE_NONENCRYPTED_FLOW_TUTORIAL_REQUEST_CODE)
+            startActivityForResult(
+                Intent(this, WelcomeActivity::class.java),
+                UPGRADE_NONENCRYPTED_FLOW_TUTORIAL_REQUEST_CODE
+            )
             overridePendingTransition(R.anim.fade_in, R.anim.fade_out)
         }
     }
 
     private fun upgradeUnencryptedWallet() {
         viewModel.finishUnecryptedWalletUpgradeAction.observe(this) {
-            startActivityForResult(SetPinActivity.createIntent(application, R.string.set_pin_upgrade_wallet, upgradingWallet = true), SET_PIN_REQUEST_CODE)
+            startActivityForResult(
+                SetPinActivity.createIntent(application, R.string.set_pin_upgrade_wallet, upgradingWallet = true),
+                SET_PIN_REQUEST_CODE
+            )
         }
         viewModel.upgradeUnencryptedWallet()
     }
@@ -174,11 +186,13 @@ class OnboardingActivity : RestoreFromFileActivity() {
     }
 
     private fun regularFlow() {
-        if (walletApplication.configuration.v7TutorialCompleted) {
+        if (config.v7TutorialCompleted) {
             upgradeOrStartMainActivity()
         } else {
-            startActivityForResult(Intent(this, WelcomeActivity::class.java),
-                    REGULAR_FLOW_TUTORIAL_REQUEST_CODE)
+            startActivityForResult(
+                Intent(this, WelcomeActivity::class.java),
+                REGULAR_FLOW_TUTORIAL_REQUEST_CODE
+            )
             overridePendingTransition(R.anim.fade_in, R.anim.fade_out)
         }
     }
@@ -192,7 +206,10 @@ class OnboardingActivity : RestoreFromFileActivity() {
     }
 
     private fun startMainActivity() {
-        startActivity(WalletActivity.createIntent(this))
+        val intent = Intent(this, WalletActivity::class.java).apply {
+            putExtras(this@OnboardingActivity.intent.extras ?: Bundle())
+        }
+        startActivity(intent)
         finish()
     }
 
@@ -200,31 +217,31 @@ class OnboardingActivity : RestoreFromFileActivity() {
         initView()
         initViewModel()
         showButtonsDelayed()
-        if (!walletApplication.configuration.v7TutorialCompleted) {
+        if (!config.v7TutorialCompleted) {
             startActivity(Intent(this, WelcomeActivity::class.java))
             overridePendingTransition(android.R.anim.fade_in, android.R.anim.fade_out)
         }
     }
 
     private fun initView() {
-        create_new_wallet.setOnClickListener {
+        binding.createNewWallet.setOnClickListener {
             viewModel.createNewWallet()
         }
-        recovery_wallet.setOnClickListener {
+        binding.recoveryWallet.setOnClickListener {
             walletApplication.initEnvironmentIfNeeded()
             startActivityForResult(Intent(this, RestoreWalletFromSeedActivity::class.java), REQUEST_CODE_RESTORE_WALLET)
         }
-        restore_wallet.setOnClickListener {
+        binding.restoreWallet.setOnClickListener {
             restoreWalletFromFile()
         }
     }
 
     @SuppressLint("StringFormatInvalid")
     private fun initViewModel() {
-        viewModel.showToastAction.observe(this, Observer {
+        viewModel.showToastAction.observe(this) {
             Toast.makeText(this, it, Toast.LENGTH_LONG).show()
-        })
-        viewModel.showRestoreWalletFailureAction.observe(this, Observer {
+        }
+        viewModel.showRestoreWalletFailureAction.observe(this) {
             val message = when {
                 TextUtils.isEmpty(it.message) -> it.javaClass.simpleName
                 else -> it.message!!
@@ -241,16 +258,19 @@ class OnboardingActivity : RestoreFromFileActivity() {
                     RestoreWalletFromSeedDialogFragment.show(supportFragmentManager)
                 }
             }
-        })
+        }
         viewModel.finishCreateNewWalletAction.observe(this) {
-            startActivityForResult(SetPinActivity.createIntent(application, R.string.set_pin_create_new_wallet), SET_PIN_REQUEST_CODE)
+            startActivityForResult(
+                SetPinActivity.createIntent(application, R.string.set_pin_create_new_wallet),
+                SET_PIN_REQUEST_CODE
+            )
         }
     }
 
     private fun showButtonsDelayed() {
-        Handler().postDelayed({
+        binding.buttons.postDelayed({
             hideSlogan()
-            findViewById<LinearLayout>(R.id.buttons).visibility = View.VISIBLE
+            binding.buttons.isVisible = true
         }, 1000)
     }
 
@@ -275,11 +295,15 @@ class OnboardingActivity : RestoreFromFileActivity() {
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
+
         if (requestCode == REGULAR_FLOW_TUTORIAL_REQUEST_CODE) {
             upgradeOrStartMainActivity()
         } else if (requestCode == UPGRADE_NONENCRYPTED_FLOW_TUTORIAL_REQUEST_CODE) {
             upgradeUnencryptedWallet()
-        } else if ((requestCode == SET_PIN_REQUEST_CODE || requestCode == RESTORE_PHRASE_REQUEST_CODE) && resultCode == Activity.RESULT_OK) {
+        } else if (
+            (requestCode == SET_PIN_REQUEST_CODE || requestCode == RESTORE_PHRASE_REQUEST_CODE) &&
+            resultCode == Activity.RESULT_OK
+        ) {
             finish()
         }
     }
