@@ -19,15 +19,13 @@ package de.schildbach.wallet.ui
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
-import androidx.core.os.bundleOf
 import dagger.hilt.android.AndroidEntryPoint
-import de.schildbach.wallet.WalletApplication
-import de.schildbach.wallet.ui.rates.ExchangeRatesFragment.*
-import de.schildbach.wallet.ui.more.AboutActivity
 import de.schildbach.wallet.ui.main.WalletActivity
+import de.schildbach.wallet.ui.more.AboutActivity
 import de.schildbach.wallet.ui.rates.ExchangeRatesActivity
+import de.schildbach.wallet.ui.rates.ExchangeRatesFragment.*
 import de.schildbach.wallet_test.R
-import kotlinx.android.synthetic.main.activity_settings.*
+import de.schildbach.wallet_test.databinding.ActivitySettingsBinding
 import org.dash.wallet.common.data.entity.ExchangeRate
 import org.dash.wallet.common.services.SystemActionsService
 import org.dash.wallet.common.services.analytics.AnalyticsConstants
@@ -36,30 +34,32 @@ import org.dash.wallet.common.ui.dialogs.AdaptiveDialog
 import org.slf4j.LoggerFactory
 import javax.inject.Inject
 
-
 @AndroidEntryPoint
-class SettingsActivity : BaseMenuActivity() {
+class SettingsActivity : LockScreenActivity() {
     companion object Constants {
         private const val RC_DEFAULT_FIAT_CURRENCY_SELECTED: Int = 100
     }
 
     private val log = LoggerFactory.getLogger(SettingsActivity::class.java)
+    private lateinit var binding: ActivitySettingsBinding
     @Inject
     lateinit var analytics: AnalyticsService
     @Inject
     lateinit var systemActions: SystemActionsService
 
-    override fun getLayoutId(): Int {
-        return R.layout.activity_settings
-    }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setTitle(R.string.settings_title)
-        about.setOnClickListener {
+
+        binding = ActivitySettingsBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        binding.appBar.toolbar.title = getString(R.string.settings_title)
+        binding.appBar.toolbar.setNavigationOnClickListener { finish() }
+
+        binding.about.setOnClickListener {
             analytics.logEvent(AnalyticsConstants.Settings.ABOUT, mapOf())
             startActivity(Intent(this, AboutActivity::class.java))
         }
-        local_currency.setOnClickListener {
+        binding.localCurrency.setOnClickListener {
             analytics.logEvent(AnalyticsConstants.Settings.LOCAL_CURRENCY, mapOf())
             val intent = Intent(this, ExchangeRatesActivity::class.java)
             intent.putExtra(ARG_SHOW_AS_DIALOG, false)
@@ -67,14 +67,13 @@ class SettingsActivity : BaseMenuActivity() {
             startActivityForResult(intent, RC_DEFAULT_FIAT_CURRENCY_SELECTED)
         }
 
-        rescan_blockchain.setOnClickListener { resetBlockchain() }
-        notifications.setOnClickListener { systemActions.openNotificationSettings() }
+        binding.rescanBlockchain.setOnClickListener { resetBlockchain() }
+        binding.notifications.setOnClickListener { systemActions.openNotificationSettings() }
     }
 
     override fun onStart() {
         super.onStart()
-        local_currency_symbol.text = WalletApplication.getInstance()
-                .configuration.exchangeCurrencyCode
+        binding.localCurrencySymbol.text = configuration.exchangeCurrencyCode
     }
 
     private fun resetBlockchain() {
@@ -89,8 +88,8 @@ class SettingsActivity : BaseMenuActivity() {
                 log.info("manually initiated blockchain reset")
                 analytics.logEvent(AnalyticsConstants.Settings.RESCAN_BLOCKCHAIN_RESET, mapOf())
 
-                WalletApplication.getInstance().resetBlockchain()
-                WalletApplication.getInstance().configuration.updateLastBlockchainResetTime()
+                walletApplication.resetBlockchain()
+                configuration.updateLastBlockchainResetTime()
                 startActivity(WalletActivity.createIntent(this@SettingsActivity))
             } else {
                 analytics.logEvent(AnalyticsConstants.Settings.RESCAN_BLOCKCHAIN_DISMISS, mapOf())
@@ -99,10 +98,9 @@ class SettingsActivity : BaseMenuActivity() {
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        if(resultCode == Activity.RESULT_OK && requestCode == RC_DEFAULT_FIAT_CURRENCY_SELECTED){
+        if (resultCode == Activity.RESULT_OK && requestCode == RC_DEFAULT_FIAT_CURRENCY_SELECTED) {
             val exchangeRate: ExchangeRate? = data?.getParcelableExtra(BUNDLE_EXCHANGE_RATE)
-            local_currency_symbol.text = if(exchangeRate != null) exchangeRate.currencyCode else
-                WalletApplication.getInstance().configuration.exchangeCurrencyCode
+            binding.localCurrencySymbol.text = exchangeRate?.currencyCode ?: configuration.exchangeCurrencyCode
         } else {
             super.onActivityResult(requestCode, resultCode, data)
         }
