@@ -19,41 +19,39 @@ package de.schildbach.wallet.ui.widget
 import android.content.Context
 import android.graphics.drawable.AnimationDrawable
 import android.util.AttributeSet
+import android.view.LayoutInflater
 import android.view.View
-import android.view.View.OnLongClickListener
 import android.widget.FrameLayout
 import androidx.annotation.StringRes
 import androidx.core.content.res.ResourcesCompat
 import de.schildbach.wallet_test.R
-import kotlinx.android.synthetic.main.payment_request_wait_error_view.view.*
-
+import de.schildbach.wallet_test.databinding.PaymentRequestWaitErrorViewBinding
 
 class PaymentRequestWaitErrorView(context: Context, attrs: AttributeSet) : FrameLayout(context, attrs) {
+    private val binding = PaymentRequestWaitErrorViewBinding.inflate(LayoutInflater.from(context), this)
 
     init {
-        inflate(context, R.layout.payment_request_wait_error_view, this)
-
         val attrsArray = context.obtainStyledAttributes(attrs, R.styleable.PaymentRequestWaitErrorView)
         try {
             when (attrsArray.getInt(R.styleable.PaymentRequestWaitErrorView_mode, -1)) {
                 0 -> {
                     setBackgroundColor(ResourcesCompat.getColor(resources, R.color.dash_blue, null))
-                    title_view.setTextColor(ResourcesCompat.getColor(resources, R.color.dash_white, null))
-                    message_view.setTextColor(ResourcesCompat.getColor(resources, R.color.dash_white, null))
-                    buttons.visibility = View.GONE
-                    close.visibility = View.GONE
+                    binding.titleView.setTextColor(ResourcesCompat.getColor(resources, R.color.dash_white, null))
+                    binding.messageView.setTextColor(ResourcesCompat.getColor(resources, R.color.dash_white, null))
+                    binding.buttons.visibility = View.GONE
+                    binding.close.visibility = View.GONE
                 }
                 1 -> {
                     setBackgroundColor(ResourcesCompat.getColor(resources, R.color.dash_white, null))
-                    title_view.setTextColor(ResourcesCompat.getColor(resources, R.color.dash_black, null))
-                    message_view.setTextColor(ResourcesCompat.getColor(resources, R.color.dash_black, null))
-                    buttons.visibility = View.VISIBLE
-                    close.visibility = View.VISIBLE
+                    binding.titleView.setTextColor(ResourcesCompat.getColor(resources, R.color.dash_black, null))
+                    binding.messageView.setTextColor(ResourcesCompat.getColor(resources, R.color.dash_black, null))
+                    binding.buttons.visibility = View.VISIBLE
+                    binding.close.visibility = View.VISIBLE
                 }
             }
             val iconSrc = attrsArray.getResourceId(R.styleable.PaymentRequestWaitErrorView_iconSrc, -1)
             if (iconSrc > -1) {
-                icon.setImageResource(iconSrc)
+                binding.icon.setImageResource(iconSrc)
             }
         } finally {
             attrsArray.recycle()
@@ -62,51 +60,51 @@ class PaymentRequestWaitErrorView(context: Context, attrs: AttributeSet) : Frame
 
     override fun onAttachedToWindow() {
         super.onAttachedToWindow()
-        val iconDrawable = icon.drawable
+        val iconDrawable = binding.icon.drawable
         if (iconDrawable is AnimationDrawable) {
             iconDrawable.start()
         }
     }
 
     fun hideConfirmButton() {
-        confirm.visibility = View.GONE
+        binding.confirm.visibility = View.GONE
     }
 
     fun setOnCloseClickListener(listener: OnClickListener) {
-        close.setOnClickListener(listener)
+        binding.close.setOnClickListener(listener)
     }
 
     fun setOnConfirmClickListener(@StringRes testResId: Int, listener: OnClickListener) {
-        confirm.setText(testResId)
-        confirm.setOnClickListener(listener)
+        binding.confirm.setText(testResId)
+        binding.confirm.setOnClickListener(listener)
     }
 
     fun setOnCancelClickListener(listener: OnClickListener) {
-        cancel.setOnClickListener(listener)
+        binding.cancel.setOnClickListener(listener)
     }
 
     fun setOnCancelClickListener(@StringRes testResId: Int, listener: OnClickListener) {
-        cancel.setText(testResId)
-        cancel.setOnClickListener(listener)
+        binding.cancel.setText(testResId)
+        binding.cancel.setOnClickListener(listener)
     }
 
     fun setMessage(message: String?) {
-        message_view.text = message
+        binding.messageView.text = message
     }
 
     var message: Int = -1
-        set(value) = message_view.setText(value)
+        set(value) = binding.messageView.setText(value)
 
     var title: Int = -1
-        set(value) = title_view.setText(value)
+        set(value) = binding.titleView.setText(value)
 
     var details: String? = null
-        set(value) = icon.setOnLongClickListener(OnLongClickListener {
-            details_view.text = value
-            if (details_view.text.isNotEmpty()) {
-                details_view.visibility = View.VISIBLE
-            }
-            return@OnLongClickListener true
-        })
+        set(value) = binding.icon.setOnLongClickListener {
+            binding.detailsView.text = value
 
+            if (binding.detailsView.text.isNotEmpty()) {
+                binding.detailsView.visibility = View.VISIBLE
+            }
+            return@setOnLongClickListener true
+        }
 }
