@@ -30,7 +30,7 @@ import androidx.fragment.app.FragmentManager
 import dagger.hilt.android.internal.managers.ViewComponentManager
 import org.dash.wallet.common.R
 import org.dash.wallet.common.databinding.ViewPaymentMethodBinding
-import org.dash.wallet.common.ui.getRoundedRippleBackground
+import org.dash.wallet.common.ui.setRoundedRippleBackground
 import org.dash.wallet.common.ui.radio_group.IconSelectMode
 import org.dash.wallet.common.ui.radio_group.IconifiedViewItem
 import org.dash.wallet.common.ui.radio_group.OptionPickerDialog
@@ -57,7 +57,7 @@ class PaymentMethodPicker(context: Context, attrs: AttributeSet): ConstraintLayo
         }
 
     init {
-        background = resources.getRoundedRippleBackground(R.style.ListViewButtonBackground)
+        setRoundedRippleBackground(R.style.ListViewButtonBackground)
         val paddingStart = resources.getDimensionPixelOffset(R.dimen.default_horizontal_padding)
         val paddingEnd = resources.getDimensionPixelOffset(R.dimen.payment_method_padding_end)
         updatePadding(left=paddingStart, right=paddingEnd)
