@@ -274,8 +274,10 @@ public class BlockchainServiceImpl extends LifecycleService implements Blockchai
                         notificationService.showNotification(
                                 "deposit_received",
                                 getString(R.string.crowdnode_deposit_received),
-                                false,
-                                new Intent(BlockchainServiceImpl.this, StakingActivity.class)
+                                null,
+                                null,
+                                new Intent(BlockchainServiceImpl.this, StakingActivity.class),
+                                null
                         );
                     } else if (apiConfirmationHandler != null && apiConfirmationHandler.matches(tx)) {
                         apiConfirmationHandler.handle(tx);
@@ -342,10 +344,9 @@ public class BlockchainServiceImpl extends LifecycleService implements Blockchai
         String msgSuffix = packageFlavor != null ? " [" + packageFlavor + "]" : "";
 
         if (exchangeRate != null) {
-            exchangeRate.coinToFiat(amount);
             MonetaryFormat format = Constants.LOCAL_FORMAT.code(0,
                     PREFIX_ALMOST_EQUAL_TO + exchangeRate.fiat.getCurrencyCode());
-            msgSuffix += " " + format.format(exchangeRate.coinToFiat(amount));
+            msgSuffix += " " + format.format(exchangeRate.coinToFiat(notificationAccumulatedAmount));
         }
 
         final String tickerMsg = getString(R.string.notification_coins_received_msg, btcFormat.format(amount))
