@@ -17,26 +17,33 @@
 package de.schildbach.wallet.ui.buy_sell
 
 import androidx.lifecycle.*
+import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import de.schildbach.wallet.data.BuyAndSellDashServicesModel
 import de.schildbach.wallet.data.ServiceStatus
 import de.schildbach.wallet.data.ServiceType
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 import org.bitcoinj.core.Coin
 import org.bitcoinj.utils.ExchangeRate
 import org.dash.wallet.common.Configuration
+import org.dash.wallet.common.WalletDataProvider
+import org.dash.wallet.common.data.ResponseResource
 import org.dash.wallet.common.services.ExchangeRatesProvider
+import org.dash.wallet.common.services.NetworkStateInt
 import org.dash.wallet.common.services.analytics.AnalyticsConstants
 import org.dash.wallet.common.services.analytics.AnalyticsService
-import org.dash.wallet.common.data.ResponseResource
-import org.dash.wallet.common.services.NetworkStateInt
 import org.dash.wallet.integration.coinbase_integration.repository.CoinBaseRepository
 import org.dash.wallet.integration.coinbase_integration.utils.CoinbaseConfig
+import org.dash.wallet.integration.uphold.api.TopperClient
 import org.dash.wallet.integration.uphold.api.UpholdClient
 import org.dash.wallet.integration.uphold.api.getDashBalance
 import org.dash.wallet.integration.uphold.api.hasValidCredentials
+import org.dash.wallet.integration.uphold.api.isAuthenticated
 import javax.inject.Inject
 
 /**
@@ -49,8 +56,10 @@ class BuyAndSellViewModel @Inject constructor(
     private val coinbaseConfig: CoinbaseConfig,
     val analytics: AnalyticsService,
     private val upholdClient: UpholdClient,
+    private val topperClient: TopperClient,
     private val networkState: NetworkStateInt,
-    exchangeRates: ExchangeRatesProvider
+    exchangeRates: ExchangeRatesProvider,
+    private val walletData: WalletDataProvider
 ): ViewModel() {
 
     companion object {
@@ -59,9 +68,8 @@ class BuyAndSellViewModel @Inject constructor(
 
     private var currentExchangeRate: org.dash.wallet.common.data.entity.ExchangeRate? = null
 
-    private val _servicesList = MutableLiveData(BuyAndSellDashServicesModel.getBuyAndSellDashServicesList())
-    val servicesList: LiveData<List<BuyAndSellDashServicesModel>>
-        get() = _servicesList
+    private val _servicesList = MutableStateFlow(BuyAndSellDashServicesModel.getBuyAndSellDashServicesList())
+    val servicesList: StateFlow<List<BuyAndSellDashServicesModel>> = _servicesList.asStateFlow()
 
     val isDeviceConnectedToInternet: LiveData<Boolean> = networkState.isConnected.asLiveData()
 
@@ -72,7 +80,9 @@ class BuyAndSellViewModel @Inject constructor(
         get() = coinBaseRepository.isAuthenticated
 
     val hasValidCredentials: Boolean
-        get() = upholdClient.hasValidCredentials
+        get() = upholdClient.hasValidCredentials &&
+            coinBaseRepository.hasValidCredentials &&
+            topperClient.hasValidCredentials
 
     init {
         exchangeRates.observeExchangeRate(config.exchangeCurrencyCode!!)
@@ -90,6 +100,10 @@ class BuyAndSellViewModel @Inject constructor(
                 updateBalances()
             }
             .launchIn(viewModelScope)
+
+        viewModelScope.launch {
+            topperClient.refreshSupportedAssets()
+        }
     }
 
     private fun setDashServiceList(list: List<BuyAndSellDashServicesModel>) {
@@ -98,7 +112,7 @@ class BuyAndSellViewModel @Inject constructor(
 
     fun updateServicesStatus() {
         setDashServiceList(
-            (_servicesList.value ?: listOf()).map { model ->
+            _servicesList.value.map { model ->
                 val serviceStatus = getItemStatus(model.serviceType)
                 if (serviceStatus != model.serviceStatus) {
                     model.copy(serviceStatus = serviceStatus)
@@ -126,6 +140,10 @@ class BuyAndSellViewModel @Inject constructor(
         var hasValidCredentials = false
 
         when (service) {
+            ServiceType.TOPPER -> {
+                hasValidCredentials = topperClient.hasValidCredentials
+                isAuthenticated = false
+            }
             ServiceType.UPHOLD -> {
                 hasValidCredentials = upholdClient.hasValidCredentials
                 isAuthenticated = upholdClient.isAuthenticated
@@ -150,7 +168,7 @@ class BuyAndSellViewModel @Inject constructor(
     }
 
     private fun showRowBalance(serviceType: ServiceType, amount: String) {
-        val list = (_servicesList.value ?: listOf()).map { model ->
+        val list = _servicesList.value.map { model ->
             if (model.serviceType == serviceType) {
                 val balance = try {
                     Coin.parseCoin(amount)
@@ -178,7 +196,9 @@ class BuyAndSellViewModel @Inject constructor(
         viewModelScope.launch {
             when (val response = coinBaseRepository.getUserAccount()) {
                 is ResponseResource.Success -> {
-                    response.value?.balance?.amount?.let { coinbaseConfig.setPreference(CoinbaseConfig.LAST_BALANCE, Coin.parseCoin(it).value) }
+                    response.value?.balance?.amount?.let {
+                        coinbaseConfig.setPreference(CoinbaseConfig.LAST_BALANCE, Coin.parseCoin(it).value)
+                    }
                     showRowBalance(ServiceType.COINBASE, response.value?.balance?.amount ?: coinbaseBalanceString())
                 }
                 is ResponseResource.Failure -> {
@@ -204,19 +224,33 @@ class BuyAndSellViewModel @Inject constructor(
     }
 
     fun logEnterUphold() {
-        analytics.logEvent(if (upholdClient.isAuthenticated) {
-            AnalyticsConstants.Uphold.ENTER_CONNECTED
-        } else {
-            AnalyticsConstants.Uphold.ENTER_DISCONNECTED
-        }, mapOf())
+        analytics.logEvent(
+            if (upholdClient.isAuthenticated) {
+                AnalyticsConstants.Uphold.ENTER_CONNECTED
+            } else {
+                AnalyticsConstants.Uphold.ENTER_DISCONNECTED
+            },
+            mapOf()
+        )
     }
 
     fun logEnterCoinbase() {
-        analytics.logEvent(if (coinBaseRepository.isUserConnected()) {
-            AnalyticsConstants.Coinbase.ENTER_CONNECTED
-        } else {
-            AnalyticsConstants.Coinbase.ENTER_DISCONNECTED
-        }, mapOf())
+        analytics.logEvent(
+            if (coinBaseRepository.isUserConnected()) {
+                AnalyticsConstants.Coinbase.ENTER_CONNECTED
+            } else {
+                AnalyticsConstants.Coinbase.ENTER_DISCONNECTED
+            },
+            mapOf()
+        )
+    }
+
+    fun topperBuyUrl(walletName: String): String {
+        return topperClient.getOnRampUrl(
+            config.exchangeCurrencyCode!!,
+            walletData.freshReceiveAddress(),
+            walletName
+        )
     }
 
     private suspend fun coinbaseBalanceString(): String =
