@@ -39,8 +39,6 @@ import de.schildbach.wallet.ui.payments.SweepWalletActivity
 import de.schildbach.wallet.util.Toast
 import de.schildbach.wallet_test.R
 import de.schildbach.wallet_test.databinding.FragmentToolsBinding
-import kotlinx.android.synthetic.main.fragment_tools.*
-import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.launch
 import org.dash.wallet.common.services.analytics.AnalyticsConstants
 import org.dash.wallet.common.services.analytics.AnalyticsService
@@ -51,7 +49,6 @@ import org.dash.wallet.common.util.Qr
 import org.slf4j.LoggerFactory
 import javax.inject.Inject
 
-@FlowPreview
 @AndroidEntryPoint
 class ToolsFragment : Fragment(R.layout.fragment_tools) {
     @Inject lateinit var authManager: SecurityFunctions
@@ -111,7 +108,7 @@ class ToolsFragment : Fragment(R.layout.fragment_tools) {
             isSyncing = it?.replaying == true
         }
 
-        transaction_export.setOnClickListener {
+        binding.transactionExport.setOnClickListener {
             if (isSyncing) {
                 val dialog = AdaptiveDialog.create(
                     null,
